refactor(app): inline fibre/valaddr wiring and remove app/fibre.go
The fibre.go helpers were seams from the removed build-tag system. With
fibre always compiled in, dissolve the thin wrappers into their single
call sites and treat fibre/valaddr as first-class modules: keepers built
inline in NewCelestiaApp, modules registered in NewManager, and added
directly to the begin/end blocker, init genesis, store key, and maccPerms
lists. Relocate countMsgPayForFibre and processFibreTxsForSquare to
filtered_square_builder.go (their natural home) and delete app/fibre.go.
