+++ app/app_test.go
@@ -120,16 +120,14 @@ func TestModuleAccountAddrs(t *testing.T) {
 			"celestia1vlthgax23ca9syk7xgaz347xmf4nunefkz88ka": true,
 			"celestia1yl6hdjhmkf37639730gffanpzndzdpmhl48edw": true,
 		}
-		for _, name := range app.FibreModuleAccountNames() {
-			want[authtypes.NewModuleAddress(name).String()] = true
-		}
+		want[authtypes.NewModuleAddress("fibre").String()] = true
 		assert.Equal(t, want, got)
 	})
 	t.Run("should be able to rederive the module account addresses from the module names", func(t *testing.T) {
 		testApp := getTestApp()
 		got := testApp.ModuleAccountAddrs()
 
-		moduleNames := append([]string{ //nolint:prealloc
+		moduleNames := []string{
 			"fee_collector",
 			"distribution",
 			"gov",
@@ -141,7 +139,8 @@ func TestModuleAccountAddrs(t *testing.T) {
 			"hyperlane",
 			"warp",
 			"forwarding",
-		}, app.FibreModuleAccountNames()...)
+			"fibre",
+		}
 		for _, moduleName := range moduleNames {
 			address := authtypes.NewModuleAddress(moduleName).String()
 			assert.Contains(t, got, address)
@@ -159,7 +158,7 @@ func TestBlockedAddresses(t *testing.T) {
 		assert.NotContains(t, got, govAddress)
 	})
 	t.Run("blocked addresses should contain all the other module addresses", func(t *testing.T) {
-		moduleNames := append([]string{ //nolint:prealloc
+		moduleNames := []string{
 			"fee_collector",
 			"distribution",
 			"mint",
@@ -170,7 +169,8 @@ func TestBlockedAddresses(t *testing.T) {
 			"hyperlane",
 			"warp",
 			"forwarding",
-		}, app.FibreModuleAccountNames()...)
+			"fibre",
+		}
 		for _, moduleName := range moduleNames {
 			address := authtypes.NewModuleAddress(moduleName).String()
 			assert.Contains(t, got, address)
